+++ CHANGELOG.md
@@ -8,9 +8,9 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## 0.9.1
 
-### Adds
+### Changes
 
--   Section modifiers for `Breadcrumbs`, including: `--locations`, `--whats-on`, `--research`, and `--books-and-more`
+-   `Breadcrumbs`'s background color now reflects the app's globally applied section modifier, such as `.nypl--locations`. It recognizes these for `--locations`, `--whats-on`, `--research`, and `--books-and-more`.
 
 ## 0.9.0